requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
modeltranslation==0.25
packaging==23.0
passlib==1.7.4
Pillow==9.5.0
PyJWT==2.6.0
pytesseract==0.3.10
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
tesseract==0.1.3
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
